// src/app/models/employee/policy/foretype.ts
export class ForetypeModel {
    constructor() {
  
    }
  
    foretype_id: string = "0";
    foretype_code: string = "";
    foretype_name_th: string = "";
    foretype_name_en: string = "";
    modified_by: string = "";
    modified_date: string = "";
    index: number = 0;
    select: boolean = false;
  
    company_code: string = "";
  }
  